🔥 Remove old testing code
